HTML, CSS, and JS for website
The Twitter links might not work because this time around I was using a local host. They should work fine on a remote server since I embedded them correctly through Twitter.

Also, the FB/Twitter buttons probably won't share correctly, because they can't link to a local host. However, Just google the Facebook share button and Twitter share button plug in and it should direct you to change the script for the plug in, which matters due to URL connection.